chore(s3): Move lib to the AWS provider
